% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_n_smallest.R
\name{pick_n_smallest}
\alias{pick_n_smallest}
\title{Getting smallest n function: picking up the n smallest value based on a given varible in a table}
\usage{
pick_n_smallest(data, col_label, n)
}
\arguments{
\item{data}{the data we focus on}

\item{col_label}{the varible we focus on in the data}

\item{n}{the number of smallest value need}
}
\value{
the new table with n rows
}
\description{
Getting smallest n function: picking up the n smallest value based on a given varible in a table
}
\examples{
library(datateachr)
library(tidyverse)
library(testthat)
library(dplyr)
library(digest)
pick_n_smallest(cancer_sample, col_label = radius_mean, 100)
}
